refactor: move log out of model, add display feature
