--- common/src/models/log.rs
@@ -1,73 +1,0 @@
-use std::fmt::Display;
-use std::fmt::Write;
-
-use chrono::{DateTime, Local};
-use crossterm::style::{StyledContent, Stylize};
-
-use crate::log::{Item, Level};
-use crate::STATE_MESSAGE;
-
-impl Display for Item {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let datetime: DateTime<Local> = DateTime::from(self.timestamp);
-
-        let message = match serde_json::from_slice(&self.fields).unwrap() {
-            serde_json::Value::String(str_value) if str_value == STATE_MESSAGE => {
-                writeln!(f)?;
-                format!("Entering {} state", self.state)
-                    .bold()
-                    .blue()
-                    .to_string()
-            }
-            serde_json::Value::Object(map) => {
-                let mut simple = None;
-                let mut extra = vec![];
-
-                for (key, value) in map.iter() {
-                    match key.as_str() {
-                        "message" => simple = value.as_str(),
-                        _ => extra.push(format!("{key}={value}")),
-                    }
-                }
-
-                let mut output = if extra.is_empty() {
-                    String::new()
-                } else {
-                    format!("{{{}}} ", extra.join(" "))
-                };
-
-                if !self.target.is_empty() {
-                    let target = format!("{}:", self.target).dim();
-                    write!(output, "{target} ")?;
-                }
-
-                if let Some(msg) = simple {
-                    write!(output, "{msg}")?;
-                }
-
-                output
-            }
-            other => other.to_string(),
-        };
-
-        write!(
-            f,
-            "{} {} {}",
-            datetime.format("%Y-%m-%dT%H:%M:%S.%fZ").to_string().dim(),
-            self.level.get_colored(),
-            message
-        )
-    }
-}
-
-impl Level {
-    fn get_colored(&self) -> StyledContent<&str> {
-        match self {
-            Level::Trace => "TRACE".magenta(),
-            Level::Debug => "DEBUG".blue(),
-            Level::Info => " INFO".green(),
-            Level::Warn => " WARN".yellow(),
-            Level::Error => "ERROR".red(),
-        }
-    }
-}
