+++ common/src/models/mod.rs
@@ -1,5 +1,4 @@
 pub mod deployment;
-pub mod log;
 pub mod project;
 pub mod resource;
 pub mod secret;
